- Win32 fixes - COM module improvements from Boris Wedl
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@15041 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/standard/basic_functions.c
@@ -3153,7 +3153,7 @@ PHP_FUNCTION(getservbyport)
 	convert_to_long_ex(port);
 	convert_to_string_ex(proto);
 
-	serv = getservbyport(htons((*port)->value.lval),(*proto)->value.str.val);
+	serv = getservbyport(htons((unsigned short) (*port)->value.lval),(*proto)->value.str.val);
 
 	if(serv == NULL)
 		RETURN_FALSE;
